refactor: store file
